Fix bug in minIO shell script deployment
The ping-minio-server init container is intended to ping the minIO server until it is ready.
However, it was found that the minio-setup init container was being run before the previous
ping-minio-server init container completed, which resulted in minio-setup failing when it
attempted to add an S3 alias. An additional minIO server ping was added to minio-setup to
ensure that the minIO server is ready before attempting to add the S3 alias and the datasets
bucket. In the future, we may consider removing the ping-minio-server init container, as it is
now redundant.

Specific release versions for the minio/minio and minio/mc docker images were also added.